refactor: merge cluster agent&dialer&manager&ops into cluster dir
