+++ cmd/cluster-manager/main.go
@@ -18,7 +18,7 @@ import (
 	"github.com/erda-project/erda-infra/base/servicehub"
 	"github.com/erda-project/erda/pkg/common"
 
-	_ "github.com/erda-project/erda/modules/cluster-manager"
+	_ "github.com/erda-project/erda/modules/cluster/cluster-manager"
 )
 
 func main() {
