+++ cmd/cluster-ops/main.go
@@ -19,7 +19,7 @@ import (
 	"github.com/erda-project/erda/pkg/common"
 
 	_ "github.com/erda-project/erda-infra/providers"
-	_ "github.com/erda-project/erda/modules/cluster-ops"
+	_ "github.com/erda-project/erda/modules/cluster/cluster-ops"
 )
 
 func main() {
