+++ modules/cluster/cluster-dialer/auth/auth.go
@@ -25,7 +25,7 @@ import (
 
 	tokenpb "github.com/erda-project/erda-proto-go/core/token/pb"
 	"github.com/erda-project/erda/apistructs"
-	"github.com/erda-project/erda/modules/cluster-dialer/config"
+	"github.com/erda-project/erda/modules/cluster/cluster-dialer/config"
 )
 
 const (
