+++ modules/cluster/cluster-dialer/server/tunnel-server.go
@@ -39,8 +39,8 @@ import (
 	tokenpb "github.com/erda-project/erda-proto-go/core/token/pb"
 	"github.com/erda-project/erda/apistructs"
 	"github.com/erda-project/erda/bundle"
-	"github.com/erda-project/erda/modules/cluster-dialer/auth"
-	"github.com/erda-project/erda/modules/cluster-dialer/config"
+	"github.com/erda-project/erda/modules/cluster/cluster-dialer/auth"
+	"github.com/erda-project/erda/modules/cluster/cluster-dialer/config"
 	"github.com/erda-project/erda/pkg/http/httputil"
 )
